ST-HS-OSX-JAM table

+++ src/Dealer.cpp
@@ -6,8 +6,8 @@
 
 #include "Dealer.hpp"
 
-std::array<int, 52> Dealer::newShuffledDeck() {
-    std::array<int, 52> newDeck = deck;
+std::array<short, 52> Dealer::newShuffledDeck() {
+    std::array<short, 52> newDeck = deck;
 
     unsigned seed = std::chrono::system_clock::now().time_since_epoch().count();
     shuffle(newDeck.begin(), newDeck.end(), std::default_random_engine(seed));
